combine DoseActor and LETActor into doseactors.py
